requirements.txt
# Core dependencies
flask>=2.0.0
numpy>=1.20.0
chromadb>=0.4.18
sentence-transformers>=2.2.2
tree-sitter>=0.20.1
pyyaml>=6.0
requests>=2.28.0

# LLM integration
langchain>=0.0.267
langchain-huggingface>=0.0.6
langchain-chroma>=0.0.1
langchain-openai>=0.0.5
openai>=1.3.0

# Optional dependencies
transformers>=4.36.0
accelerate>=0.21.0
ollama>=0.1.0

# Development dependencies
pytest>=7.0.0
black>=23.1.0
isort>=5.12.0
flake8>=6.0.0

# Add to existing requirements.txt
python-dotenv>=1.0.0